client util defined
