+++ src/main.rs
@@ -1,5 +1,6 @@
 use redis_starter_rust::{
     adapters::cancellation_token::CancellationTokenFactory,
+    client_utils::ClientStreamHandler,
     services::config::{actor::ConfigActor, manager::ConfigManager},
     StartUpFacade,
 };
@@ -12,4 +13,66 @@ async fn main() -> anyhow::Result<()> {
     let mut start_up_runner = StartUpFacade::new(CancellationTokenFactory, config_manager);
 
     start_up_runner.run(()).await
+}
+
+#[tokio::test]
+async fn test_cancellation_token() {
+    use redis_starter_rust::services::interface::TCancellationNotifier;
+    use redis_starter_rust::services::interface::TCancellationTokenFactory;
+    use redis_starter_rust::services::interface::TCancellationWatcher;
+    use redis_starter_rust::TNotifyStartUp;
+    use std::sync::Arc;
+
+    use tokio::net::TcpStream;
+    // GIVEN
+
+    pub struct StartFlag(pub Arc<tokio::sync::Notify>);
+    impl TNotifyStartUp for StartFlag {
+        fn notify_startup(&self) {
+            self.0.notify_waiters();
+        }
+    }
+    struct TestNotifier(tokio::sync::oneshot::Sender<()>);
+    impl TCancellationNotifier for TestNotifier {
+        fn notify(self) {
+            let _ = self.0.send(());
+        }
+    }
+    #[derive(Clone, Copy)]
+    pub struct TestCancellationTokenFactory;
+    impl TCancellationTokenFactory for TestCancellationTokenFactory {
+        fn create(&self, _timeout: u64) -> (impl TCancellationNotifier, impl TCancellationWatcher) {
+            let (tx, rx) = tokio::sync::oneshot::channel();
+
+            (TestNotifier(tx), rx)
+        }
+    }
+
+    let config = ConfigActor::default();
+    let mut manager = ConfigManager::new(config);
+
+    manager.port = 55534;
+
+    let notify = Arc::new(tokio::sync::Notify::new());
+    let start_flag = StartFlag(notify.clone());
+
+    let mut start_up_facade = StartUpFacade::new(TestCancellationTokenFactory, manager.clone());
+
+    let handler = tokio::spawn(async move {
+        start_up_facade.run(start_flag).await.unwrap();
+    });
+
+    //warm up time
+    notify.notified().await;
+
+    let mut client_stream = TcpStream::connect(manager.bind_addr()).await.unwrap();
+    let mut h: ClientStreamHandler = client_stream.split().into();
+
+    // WHEN
+    h.send(b"*5\r\n$3\r\nSET\r\n$10\r\nsomanyrand\r\n$3\r\nbar\r\n$2\r\npx\r\n$3\r\n300\r\n").await;
+
+    // THEN
+    assert_eq!(h.get_response().await, "-Error operation cancelled due to timeout\r\n");
+
+    handler.abort();
 }
