+++ tests/test_keys.rs
@@ -1,20 +1,16 @@
 mod common;
-use common::{array, init_config_manager_with_free_port, start_test_server, TestStreamHandler};
+use common::{array, spawn_server_process, terminate_process};
 
-use redis_starter_rust::{
-    adapters::cancellation_token::CancellationTokenFactory, services::query_io::QueryIO,
-};
+use redis_starter_rust::{client_utils::ClientStreamHandler, services::query_io::QueryIO};
 use tokio::net::TcpStream;
 
 #[tokio::test]
 async fn test_keys() {
     // GIVEN
-    let config = init_config_manager_with_free_port().await;
+    let master_port = spawn_server_process();
 
-    let _ = start_test_server(CancellationTokenFactory, config.clone()).await;
-
-    let mut stream = TcpStream::connect(config.bind_addr()).await.unwrap();
-    let mut h: TestStreamHandler = stream.split().into();
+    let mut stream = TcpStream::connect(format!("localhost:{}", master_port)).await.unwrap();
+    let mut h: ClientStreamHandler = stream.split().into();
     let num_of_keys = 500;
 
     // WHEN set 100000 keys
@@ -28,4 +24,6 @@ async fn test_keys() {
     let res = h.get_response().await;
 
     assert!(res.starts_with(format!("*{}\r\n", num_of_keys).as_str()));
+
+    terminate_process(master_port);
 }
