+++ tests/test_save_read_dump.rs
@@ -4,16 +4,18 @@
 /// if the value of dir is /tmp, then the expected response to CONFIG GET dir is:
 /// *2\r\n$3\r\ndir\r\n$4\r\n/tmp\r\n
 mod common;
-use crate::common::{array, start_test_server};
-use common::{init_config_manager_with_free_port, TestStreamHandler};
-use redis_starter_rust::{
-    adapters::cancellation_token::CancellationTokenFactory,
-    services::{
-        config::{ConfigCommand, ConfigMessage},
-        query_io::QueryIO,
-    },
-};
-use std::time::{SystemTime, UNIX_EPOCH};
+use crate::common::array;
+use common::get_available_port;
+use common::terminate_process;
+use common::wait_for_message;
+
+use common::TestProcessChild;
+use redis_starter_rust::client_utils::ClientStreamHandler;
+use redis_starter_rust::services::query_io::QueryIO;
+use std::process::Command;
+use std::process::Stdio;
+use std::time::SystemTime;
+use std::time::UNIX_EPOCH;
 use tokio::net::TcpStream;
 
 struct FileName(String);
@@ -23,21 +25,37 @@ impl Drop for FileName {
     }
 }
 
+fn run_server_with_dbfilename(dbfilename: &str) -> u16 {
+    let port = get_available_port();
+    let mut command = Command::new("cargo");
+    command.args(["run", "--", "--port", &port.to_string(), "--dbfilename", dbfilename]);
+
+    let mut process = TestProcessChild(
+        command
+            .stdout(Stdio::piped())
+            .stderr(Stdio::piped())
+            .spawn()
+            .expect("Failed to start server process"),
+    );
+    wait_for_message(
+        process.0.stdout.as_mut().unwrap(),
+        format!("listening peer connection on localhost:{}...", port + 10000).as_str(),
+        1,
+    );
+
+    port
+}
+
 // TODO response cannot be deterministic!
 #[tokio::test]
 async fn test_save_read_dump() {
     // GIVEN
     let test_file_name = FileName(create_unique_file_name("test_save_dump"));
-    let config = init_config_manager_with_free_port().await;
-    config
-        .send(ConfigMessage::Command(ConfigCommand::SetDbFileName(test_file_name.0.clone())))
-        .await
-        .unwrap();
 
-    let _ = start_test_server(CancellationTokenFactory, config.clone()).await;
+    let master_port = run_server_with_dbfilename(test_file_name.0.as_str());
 
-    let mut client_stream = TcpStream::connect(config.bind_addr()).await.unwrap();
-    let mut h: TestStreamHandler = client_stream.split().into();
+    let mut client_stream = TcpStream::connect(format!("localhost:{}", master_port)).await.unwrap();
+    let mut h: ClientStreamHandler = client_stream.split().into();
 
     // WHEN
     // set without expiry time
@@ -59,15 +77,11 @@ async fn test_save_read_dump() {
     // wait for the file to be created
     tokio::time::sleep(tokio::time::Duration::from_secs(1)).await;
 
-    // start another instance
-    let _ = start_test_server(CancellationTokenFactory, config.clone()).await;
-
-    let mut client_stream = TcpStream::connect(config.bind_addr()).await.unwrap();
-    let mut h: TestStreamHandler = client_stream.split().into();
-
     // keys
     h.send({ array(vec!["KEYS", "*"]).into_bytes() }.as_slice()).await;
     assert_eq!(h.get_response().await, array(vec!["foo2", "foo"]));
+
+    terminate_process(master_port);
 }
 
 fn create_unique_file_name(function_name: &str) -> String {
